% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-hxl.R
\name{as_hxl}
\alias{as_hxl}
\alias{as_hxl.data.frame}
\title{Converts an object to an hxl data_frame}
\usage{
as_hxl(x)

\method{as_hxl}{data.frame}(x)
}
\arguments{
\item{x}{the object}
}
\value{
a hxl_tbl data_frame
}
\description{
Converts an object to an hxl data_frame
}

