% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{hxl_schema}
\alias{hxl_schema}
\title{Get the HXL schema of an HXL tibble}
\usage{
hxl_schema(x)
}
\arguments{
\item{x}{an HXL tibble}
}
\value{
a tibble data_frame of the schema
}
\description{
Get the HXL schema of an HXL tibble
}

