% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{hxl_select}
\alias{hxl_select}
\title{Select a subset of columns by tags/attributes}
\usage{
hxl_select(hxl, tags)
}
\arguments{
\item{hxl}{an HXL table}

\item{tags}{a character vector of HXL tags

It warns if a tag matches more columns.}
}
\value{
a data.frame in the order of the tags
}
\description{
Select a subset of columns by tags/attributes
}

