% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{is.hxl}
\alias{is.hxl}
\alias{is_hxl}
\title{Check if an object is an HXL tibble}
\usage{
is.hxl(x)

is_hxl(x)
}
\arguments{
\item{x}{an object to test}
}
\value{
true iff x is an HXL tibble
}
\description{
Check if an object is an HXL tibble
}

