% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_chr}
\alias{schema_chr}
\title{Get the HXL schema of an HXL tibble as character}
\usage{
schema_chr(x)
}
\arguments{
\item{x}{an HXL tibble}
}
\value{
a character in the same order as the columns
}
\description{
Get the HXL schema of an HXL tibble as character
}

